requirements.txt
transformers
langchain
accelerate
bitsandbytes
beautifulsoup4
openai
tiktoken
sentence_transformers
InstructorEmbedding
pytesseract
pdf2image
llama-cpp-python
unstructured
python-dotenv
faiss-cpu